表结构修改，重新生成mapper

+++ common/src/main/java/com/github/tanhao1410/thesis/common/mapper/AlarmDOMapper.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 /**
  * 框架自动生成表模型和CRUD操作，勿修改；
  * 如特殊需要，请以Ext***Mapper自行扩展；
- * 生成日期 : 2021-03-01 16:27:59
+ * 生成日期 : 2021-03-02 13:25:30
  * @author ##tanhao##
  */
 @Repository
@@ -44,7 +44,7 @@ public interface AlarmDOMapper {
     /**
      * 插入行记录返回影响行数，使用getId()获取插入行记录的Id
      */
-    Integer insertSelectiveReturnPrimaryKey(AlarmDO record);
+    Long insertSelectiveReturnPrimaryKey(AlarmDO record);
 
     /**
      * 批量插入行数据返回插入行数
